#30 remove CommonComposables god class and split code used in small files. Also use external ComposeViews to replace SimpleSpace and DropdownSelector.
