+++ coreUi/src/main/java/pt/joasvpereira/coreui/counter/IconData.kt
@@ -1,0 +1,5 @@
+package pt.joasvpereira.coreui.counter
+
+import androidx.compose.ui.graphics.painter.Painter
+
+data class IconData(val painter: Painter, val contentDescription: String = "")
