+++ coreUi/src/main/java/pt/joasvpereira/coreui/dialog/Dialogs.kt
@@ -51,7 +51,7 @@ import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
 import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.constraintlayout.compose.Dimension
-import com.joasvpereira.dev.mokeupui.compose.screen.organizer.main.SimpleSpace
+import com.joasvpereira.lib.compose.spacer.SimpleSpace
 import pt.joasvpereira.coreui.theme.DynamicTheme
 
 const val DISABLED_BUTTONS_ALPHA_VALUE = .25f
