+++ main/src/main/java/com/joasvpereira/main/compose/details/ParentDetailsSection.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.joasvpereira.dev.mokeupui.compose.screen.organizer.main.SimpleSpace
+import com.joasvpereira.lib.compose.spacer.SimpleSpace
 import com.joasvpereira.main.domain.data.ItemDetail
 import com.joasvpereira.main.presentation.icons.DivisionIcons
 import pt.joasvpereira.coreui.box.BoxImage
