+++ main/src/main/java/com/joasvpereira/main/compose/division/DivisionHeader.kt
@@ -28,7 +28,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import com.joasvpereira.dev.mokeupui.compose.screen.organizer.main.SimpleSpace
+import com.joasvpereira.lib.compose.spacer.SimpleSpace
 import com.joasvpereira.main.presentation.icons.DivisionIcons
 import pt.joasvpereira.coreui.theme.DynamicTheme
 import kotlin.math.min
