+++ main/src/main/java/com/joasvpereira/main/compose/division/list/elements/BoxItem.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import com.joasvpereira.dev.mokeupui.compose.screen.organizer.main.SimpleSpace
+import com.joasvpereira.lib.compose.spacer.SimpleSpace
 import com.joasvpereira.main.compose.division.BaseDivisionItemContainer
 import pt.joasvpereira.coreui.box.BoxImage
 import pt.joasvpereira.coreui.preview.ThemesProvider
