+++ settings/src/main/java/com/joasvpereira/settings/compose/main/SettingMainView.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import com.joasvpereira.dev.mokeupui.compose.screen.organizer.main.SimpleSpace
+import com.joasvpereira.lib.compose.spacer.SimpleSpace
 import com.joasvpereira.settings.compose.main.session.SessionSettingsSection
 import com.joasvpereira.settings.compose.main.theme.ThemeSettingsSection
 import pt.joasvpereira.core.domain.data.SessionItem
