+++ settings/src/main/java/com/joasvpereira/settings/compose/main/session/SettingsSessionIndicator.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.core.graphics.drawable.toBitmap
-import com.joasvpereira.dev.mokeupui.compose.screen.organizer.main.SimpleSpace
+import com.joasvpereira.lib.compose.spacer.SimpleSpace
 import pt.joasvpereira.core.domain.data.SessionItem
 import pt.joasvpereira.coreui.R
 import pt.joasvpereira.coreui.theme.DynamicTheme
